Add unarchive based copy method to upload_helm_chart role
Synchronize module does not work when for some reason there is no
ssh to the target, which can happen when using custom ansible
connection plugins to containerised targets.

This patch adds the multi_copy action plugin which is a wrapper
around the unarchive module, copying many files in one task.